test(e2e): extract protofleet spec helpers
